Deleted mapdl module (to be impemented in a separate repo) Put all graph related code in a single file accessible from root Made signal PEP8 compliant
