--- pymodal/mapdl/__init__.py
@@ -1,44 +1,0 @@
-from .preprocessor import (  # noqa F401
-    _get_closest_node,
-    _get_max_param_id,
-    set_linear_elastic,
-    create_line,
-    create_area,
-    create_prism,
-    create_cylinder,
-    create_extruded_volume,
-    set_mass21,
-    set_link180,
-    set_beam3,
-    set_shell181,
-    set_solid186,
-    get_node_list,
-    select_nodes,
-    select_areas,
-    displacement_bc,
-    elastic_support,
-    linear_elastic_surface_contact,
-    move_volume,
-)
-
-from .solver import (
-    modal_analysis,
-    get_stiffness,
-    harmonic_analysis,
-)
-
-from .mesher import (
-    free_beam,
-    free_plate_shell,
-    free_plate_solid,
-    cantilever_beam,
-    circ_hole_solid,
-    crack_analogy_solid,
-    stringer_support_solid,
-    los_alamos_building,
-    add_spring,
-    free_beam_solid,
-    damaged_beam_solid,
-    mass_beam_solid,
-    stringer_beam_solid,
-)
