--- pymodal/papergraph/__init__.py
@@ -1,10 +1,0 @@
-from .utils import (  # noqa: F401
-    multiple_formatter,
-    add_cbar
-)
-
-from .lineplot import lineplot  # noqa: F401
-
-from .imgplot import imgplot  # noqa: F401
-
-from .surfplot import surfplot  # noqa: F401
