+++ pymodal/signal_class.py
@@ -19,7 +19,9 @@ class signal():
                  orientation: list[tuple[float, float, float]] = None,
                  units: str = "mm/s^2"):
 
-        self.amplitude = np.array(amplitude).reshape((self.amplitude.shape[0], self.amplitude.shape[1], -1))
+        self.amplitude = np.array(amplitude).reshape((
+            self.amplitude.shape[0], self.amplitude.shape[1], -1
+        ))
         self.units = units
         self.samples = self.amplitude.shape[0]
         self.measurements = self.amplitude.shape[2]
@@ -33,8 +35,9 @@ class signal():
         self.min_time = min_time  # Minimum time is assumed to be 0
         if elapsed_time is None:
             if max_time is None:
-                # If neither elapsed_time nor maximum time are defined
-                self.sample_rate = sample_rate  # Then sample_rate MUST be defined
+                # If neither elapsed_time nor maximum time are defined then
+                # sample_rate MUST be defined
+                self.sample_rate = sample_rate
                 # So elapsed_time and maximum time can be calculated
                 self.elapsed_time = (self.samples - 1) * self.sample_rate
                 self.max_time = self.min_time + self.elapsed_time
@@ -71,15 +74,17 @@ class signal():
         elapsed_time_error = not(self.elapsed_time == calculated_elapsed_time)
         if sample_rate_error or elapsed_time_error:
             raise Exception((
-                f"The sample_rate ({sample_rate} Hz), elapsed_time ({elapsed_time} s)"
-                f", min_time ({min_time} s) and/or max_time ({max_time} s)"
-                f" values introduced were not coherent with each other.\n\n"
-                f"elapsed_time = (data_points - 1) * sample_rate\n"
-                f"max_time = min_time + elapsed_time"
+                f"The sample_rate ({sample_rate} Hz), elapsed_time"
+                f" ({elapsed_time} s), min_time ({min_time} s) and/or max_time"
+                f" ({max_time} s) values introduced were not coherent with"
+                f" each other.\n\nelapsed_time = (data_points - 1) *"
+                f" sample_rate\nmax_time = min_time + elapsed_time"
             ))
         
         self.time_vector = np.arange(
-            self.min_time, self.max_time + self.sample_rate / 2, self.sample_rate
+            self.min_time,
+            self.max_time + self.sample_rate / 2,
+            self.sample_rate
         )
 
         self.coordinates = coordinates
@@ -88,7 +93,9 @@ class signal():
         else:
             for element in coordinates:
                 if len(element) != 3:
-                    raise Exception("At least one coordinate has too many elements.")
+                    raise Exception(
+                        "At least one coordinate has too many elements."
+                    )
                 if len(self.coordinates) > len(self):
                     raise Exception("Too many coordinates were provided.")
                 elif len(self.coordinates) < len(self):
@@ -103,7 +110,9 @@ class signal():
         else:
             for element in coordinates:
                 if len(element) != 3:
-                    raise Exception("At least one orientation has too many elements.")
+                    raise Exception(
+                        "At least one orientation has too many elements."
+                    )
                 if len(self.orientation) > len(self):
                     raise Exception("Too many orientations were provided.")
 
@@ -146,15 +155,17 @@ class signal():
     def __getitem__(self, index: slice):
         if not isinstance(index, slice):
             index = slice(index, index + 1, None)
-        return signal(amplitude = self.amplitude[:, index.start:index.stop:index.step],
-                      sample_rate = self.sample_rate,
-                      elapsed_time = self.elapsed_time,
-                      max_time = self.max_time,
-                      min_time = self.min_time,
-                      label = self.label[index.start:index.stop:index.step],
-                      coordinates = self.coordinates[index.start:index.stop:index.step],
-                      orientation = self.orientation[index.start:index.stop:index.step],
-                      units = self.units)
+        return signal(
+            amplitude = self.amplitude[:, index.start:index.stop:index.step],
+            sample_rate = self.sample_rate,
+            elapsed_time = self.elapsed_time,
+            max_time = self.max_time,
+            min_time = self.min_time,
+            label = self.label[index.start:index.stop:index.step],
+            coordinates = self.coordinates[index.start:index.stop:index.step],
+            orientation = self.orientation[index.start:index.stop:index.step],
+            units = self.units
+        )
 
 
     def extend(self, new_signal):
@@ -162,21 +173,24 @@ class signal():
         assert self.min_time == new_signal.min_time
         assert self.max_time == new_signal.max_time
         assert self.units == new_signal.units
-        return signal(amplitude = np.vstack((self.amplitude, new_signal.amplitude)),
-                      sample_rate = self.sample_rate,
-                      elapsed_time = self.elapsed_time,
-                      max_time = self.max_time,
-                      min_time = self.min_time,
-                      label = np.vstack((self.label, new_signal.label)),
-                      coordinates = np.vstack((self.coordinates, new_signal.coordinates)),
-                      orientation = np.vstack((self.amplitude, new_signal.amplitude)),
-                      units = self.units)
+        return signal(
+            amplitude = np.vstack((self.amplitude, new_signal.amplitude)),
+            sample_rate = self.sample_rate,
+            elapsed_time = self.elapsed_time,
+            max_time = self.max_time,
+            min_time = self.min_time,
+            label = np.vstack((self.label, new_signal.label)),
+            coordinates = np.vstack((self.coordinates,
+                                     new_signal.coordinates)),
+            orientation = np.vstack((self.amplitude, new_signal.amplitude)),
+            units = self.units
+        )
 
     
     def change_sample_rate(self, new_sample_rate: float):
         if new_sample_rate % self.sample_rate != 0:
-            warn("The resulting signal will be interpolated according to the desired new"
-                 " sample_rate.")
+            warn("The resulting signal will be interpolated according to the"
+                 " desired new sample_rate.")
             new_time_vector = np.arange(
                 self.min_freq, self.max_freq+new_sample_rate, new_sample_rate
             )
@@ -201,18 +215,25 @@ class signal():
                       units=self.units)
 
             
-    def change_sample_rate(self, new_min_time: float = None, new_max_time: float = None):
+    def change_sample_rate(self, new_min_time: float = None,
+                           new_max_time: float = None):
         if new_min_time is None:
             new_min_time = self.min_time
         if new_max_time is None:
             new_max_time = self.max_time
         new_time_vector = self.time_vector
-        closest_max_time_index = (np.abs(new_time_vector - new_max_time)).argmin()
+        closest_max_time_index = (
+            np.abs(new_time_vector - new_max_time)
+        ).argmin()
         closest_max_time = new_time_vector[closest_max_time_index]
-        closest_min_time_index = (np.abs(new_time_vector - new_min_time)).argmin()
+        closest_min_time_index = (
+            np.abs(new_time_vector - new_min_time)
+        ).argmin()
         closest_min_time = new_time_vector[closest_min_time_index]
         new_elapsed_time = closest_max_time - closest_min_time
-        new_sample_rate = new_elapsed_time / (closest_max_time_index-closest_min_time_index)
+        new_sample_rate = (
+            new_elapsed_time / (closest_max_time_index-closest_min_time_index)
+        )
         if new_sample_rate != self.sample_rate:
             new_signal = self.change_sample_rate(new_sample_rate)
         else:
@@ -221,31 +242,48 @@ class signal():
 
         if new_max_time > new_signal.max_time:
             time_extension = np.arange(
-                new_signal.max_time, new_max_time+new_signal.sample_rate, new_signal.sample_rate
+                new_signal.max_time,
+                new_max_time+new_signal.sample_rate,
+                new_signal.sample_rate
             )
             new_time_vector = np.hstack((new_time_vector, time_extension))
             amplitude_extension = np.zeros((
                 time_extension.shape[0], new_signal.amplitude.shape[-1]
             ))
-            new_amplitude = np.hstack((new_signal.amplitude, amplitude_extension))
+            new_amplitude = np.hstack((
+                new_signal.amplitude, amplitude_extension
+            ))
         else:
-            new_time_vector = new_time_vector[0:(np.abs(new_time_vector - new_max_time)).argmin()]
-            new_amplitude = new_amplitude[0:(np.abs(new_time_vector - new_max_time)).argmin(), :]
+            new_time_vector = new_time_vector[
+                0:(np.abs(new_time_vector - new_max_time)).argmin()
+            ]
+            new_amplitude = new_amplitude[
+                0:(np.abs(new_time_vector - new_max_time)).argmin(), :
+            ]
 
         if new_min_time < new_signal.min_time:
             time_extension = np.arange(
-                new_min_time, new_signal.min_time+new_signal.sample_rate, new_signal.sample_rate
+                new_min_time,
+                new_signal.min_time+new_signal.sample_rate,
+                new_signal.sample_rate
             )
             new_time_vector = np.hstack((time_extension, new_time_vector))
             amplitude_extension = np.zeros((
                 time_extension.shape[0], new_signal.amplitude.shape[-1]
             ))
-            new_amplitude = np.hstack((amplitude_extension, new_signal.amplitude))
+            new_amplitude = np.hstack((
+                amplitude_extension,
+                new_signal.amplitude
+            ))
             if new_min_time < 0:
                 new_time_vector = new_time_vector + abs(new_min_time)
         else:
-            new_time_vector = new_time_vector[(np.abs(new_time_vector - new_max_time)).argmin():]
-            new_amplitude = new_amplitude[(np.abs(new_time_vector - new_max_time)).argmin():, :]
+            new_time_vector = new_time_vector[
+                (np.abs(new_time_vector - new_max_time)).argmin():
+            ]
+            new_amplitude = new_amplitude[
+                (np.abs(new_time_vector - new_max_time)).argmin():, :
+            ]
         return signal(signal=new_amplitude,
                       sample_rate=new_sample_rate,
                       elapsed_time=max(new_time_vector)-min(new_time_vector),
@@ -260,9 +298,13 @@ class signal():
     def apply_window(self, window: str):
         if isinstance(window, str):
             if window == "hanning":
-                self.amplitude = self.amplitude * np.tile(np.hanning(self.samples), (1, len(self)))
+                self.amplitude = self.amplitude * np.tile(
+                    np.hanning(self.samples), (1, len(self))
+                )
             elif window == "hamming":
-                self.amplitude = self.amplitude * np.tile(np.hamming(self.samples), (1, len(self)))
+                self.amplitude = self.amplitude * np.tile(
+                    np.hamming(self.samples), (1, len(self))
+                )
         else:
             window = interpolate.interp1d(self.time_vector, window)
             self.amplitude = self.amplitude * np.tile(window, (1, len(self)))
